+++ packages/gatsby-remark-code-repls/src/__tests__/index.js
@@ -24,6 +24,7 @@ const REMARK_TESTS = {
 }
 
 const remark = new Remark()
+const JSONstringifySpy = jest.spyOn(JSON, `stringify`)
 
 describe(`gatsby-remark-code-repls`, () => {
   beforeEach(() => {
@@ -32,6 +33,7 @@ describe(`gatsby-remark-code-repls`, () => {
 
     fs.readFileSync.mockReset()
     fs.readFileSync.mockReturnValue(`const foo = "bar";`)
+    JSONstringifySpy.mockClear()
   })
 
   Object.keys(REMARK_TESTS).forEach(name => {
@@ -164,11 +166,32 @@ describe(`gatsby-remark-code-repls`, () => {
           const transformed = plugin(
             { markdownAST },
             {
-              dependencies: [`react`, `react-dom@next`, `prop-types@15.5`],
+              dependencies: [
+                `react`,
+                `react-dom@next`,
+                `prop-types@15.5`,
+                `@babel/core@7.4.0`,
+              ],
               directory: `examples`,
             }
           )
 
+          expect(JSONstringifySpy).toHaveBeenCalledWith(
+            expect.objectContaining({
+              files: expect.objectContaining({
+                "package.json": expect.objectContaining({
+                  content: expect.objectContaining({
+                    dependencies: expect.objectContaining({
+                      react: `latest`,
+                      "react-dom": `next`,
+                      "prop-types": `15.5`,
+                      "@babel/core": `7.4.0`,
+                    }),
+                  }),
+                }),
+              }),
+            })
+          )
           expect(transformed).toMatchSnapshot()
         })