+++ www/src/components/evaluation-table-section-header-top.js
@@ -1,5 +1,12 @@
 import React from "react"
-import presets, { colors, space, radii } from "../utils/presets"
+import {
+  colors,
+  space,
+  radii,
+  breakpoints,
+  scale,
+  lineHeights,
+} from "../utils/presets"
 import { rhythm } from "../utils/typography"
 
 const superHeaderTitles = [
@@ -17,20 +24,17 @@ const superHeader = () => (
         key={i}
         css={{
           "&&": {
-            padding: `${rhythm(space[3])} ${rhythm(presets.space[3])} ${rhythm(
-              3 / 8
-            )}`,
+            padding: `${space[3]} ${space[3]} ${rhythm(3 / 8)}`,
           },
           display: `none`,
           textTransform: `uppercase`,
-          fontSize: presets.scale[0],
-          lineHeight: presets.lineHeights.solid,
+          fontSize: scale[0],
+          lineHeight: lineHeights.solid,
           fontWeight: 500,
           textAlign: `center`,
           verticalAlign: `bottom`,
           width: i === 0 ? 120 : `inherit`,
           border: 0,
-          // fontFamily: options.headerFontFamily.join(`,`),
           color: colors.gray.calm,
           background: colors.ui.whisper,
           "span:first-of-type": {
@@ -40,14 +44,14 @@ const superHeader = () => (
           "&:last-child": {
             borderTopRightRadius: radii[2],
           },
-          [presets.Xs]: {
+          [breakpoints.xs]: {
             display: `table-cell`,
             width: 125,
           },
-          [presets.Md]: {
+          [breakpoints.md]: {
             width: 150,
           },
-          [presets.Lg]: {
+          [breakpoints.lg]: {
             width: 175,
           },
         }}