+++ www/src/views/creators/index.js
@@ -1,12 +1,11 @@
 import React, { Component } from "react"
 import { Helmet } from "react-helmet"
-import { rhythm } from "../../utils/typography"
 import Layout from "../../components/layout"
 import CreatorsHeader from "./creators-header"
 import Badge from "./badge"
 import GithubIcon from "react-icons/lib/go/mark-github"
 import { navigate } from "gatsby"
-import presets, { colors, space } from "../../utils/presets"
+import { colors, space, breakpoints, scale } from "../../utils/presets"
 import qs from "qs"
 import ThumbnailLink from "../shared/thumbnail"
 import EmptyGridItems from "../shared/empty-grid-items"
@@ -109,10 +108,10 @@ class CreatorsView extends Component {
         <main
           id={`reach-skip-nav`}
           css={{
-            padding: rhythm(space[6]),
+            padding: space[6],
             paddingBottom: `10vh`,
-            [presets.Md]: {
-              paddingBottom: rhythm(space[6]),
+            [breakpoints.md]: {
+              paddingBottom: space[6],
             },
           }}
         >
@@ -121,7 +120,7 @@ class CreatorsView extends Component {
               display: `flex`,
               flexWrap: `wrap`,
               justifyContent: `center`,
-              [presets.Lg]: {
+              [breakpoints.lg]: {
                 justifyContent: `flex-start`,
               },
             }}
@@ -141,7 +140,7 @@ class CreatorsView extends Component {
                   <div css={{ display: `flex`, ...sharedStyles.meta }}>
                     <div
                       css={{
-                        margin: `0 0 ${rhythm(space[1])}`,
+                        margin: `0 0 ${space[1]}`,
                         color: colors.gray.calm,
                       }}
                     >
@@ -163,7 +162,7 @@ class CreatorsView extends Component {
                     <div
                       css={{
                         alignSelf: `flex-start`,
-                        fontSize: presets.scale[0],
+                        fontSize: scale[0],
                       }}
                     >
                       <Badge forHire={item.node.for_hire}>
@@ -188,7 +187,7 @@ const styles = {
   creatorCard: {
     display: `flex`,
     flexDirection: `column`,
-    margin: rhythm(space[6]),
+    margin: space[6],
     minWidth: 200,
     maxWidth: 240,
     flex: `1 0 0`,