+++ packages/gatsby-codemods/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.9...gatsby-codemods@1.0.10) (2019-03-15)
+
+**Note:** Version bump only for package gatsby-codemods
+
 ## [1.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.8...gatsby-codemods@1.0.9) (2019-03-11)
 
 **Note:** Version bump only for package gatsby-codemods