+++ packages/gatsby-remark-copy-linked-files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-copy-linked-files/compare/gatsby-remark-copy-linked-files@2.0.10...gatsby-remark-copy-linked-files@2.0.11) (2019-03-15)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
 ## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-copy-linked-files/compare/gatsby-remark-copy-linked-files@2.0.9...gatsby-remark-copy-linked-files@2.0.10) (2019-03-11)
 
 **Note:** Version bump only for package gatsby-remark-copy-linked-files